refactor(flui-painting): Mythos Step 7 -- split text_painter.rs god module
Split the 990-LOC `crates/flui-painting/src/text_painter.rs` god
module into a concern-based `text_painter/` directory. Four files,
each focused on one concern; total LOC drops to ~873 (net -117 from
removing duplicate docstring noise + concise per-file headers).

Files created:

| File                            | LOC | Concern                                                  |
|---------------------------------|-----|----------------------------------------------------------|
| `text_painter/mod.rs`           | 513 | TextPainter struct + TextLayoutCache + LayoutMetrics + Default + new + builder API + getters + setters + mark_needs_layout/did_layout + inline tests |
| `text_painter/baseline.rs`      |  15 | `TextBaseline` enum (standalone type)                    |
| `text_painter/measure.rs`       | 206 | layout + compute_layout_metrics + compute_paint_offset + size/width/height + compute_distance_to_actual_baseline + did_exceed_max_lines |
| `text_painter/paint.rs`         | 139 | paint + cursor methods (get_offset_for_caret, get_position_for_offset, get_line_metrics, get_boxes_for_selection, get_word_boundary) |

The split honors concerns by responsibility:

- mod.rs owns the `TextPainter` state machine (struct + cache + 11
  builder methods + 9 getters + 9 setters + lifecycle). This is
  where mutation happens.
- baseline.rs owns the `TextBaseline` enum -- a small standalone
  type with `Default` (Alphabetic) + `serde` opt-in.
- measure.rs owns the layout-and-metrics work: `layout()` invokes
  cosmic-text via `compute_layout_metrics`, populates `TextLayoutCache`,
  exposes geometry queries (`size`/`width`/`height`/baseline).
- paint.rs owns the canvas integration: `paint()` emits the
  `DrawTextSpan` command to the canvas, plus the cursor + selection
  query helpers that read from the cached layout.

Struct fields are `pub(super)` to let `measure` and `paint`
submodules read/write the cache without `pub` exposure.

The 12 inline tests from the old `text_painter.rs` are preserved in
`text_painter/mod.rs` for this step. U8 of the chain extracts them
to `crates/flui-painting/tests/text_painter_basic.rs`.

Public API surface unchanged. External callers use the same
`flui_painting::{TextPainter, TextBaseline, DEFAULT_FONT_SIZE}`
paths.

Verification:
- `cargo check -p flui-painting` clean.
- `cargo test -p flui-painting --lib` 64 passing.
- `cargo test -p flui-painting --tests` 9 passing.
- `cargo check --workspace` clean.
- `bash scripts/port-check.sh` 6 triggers ok.

Net unsafe delta: 0.
Net LOC delta in production code: -117 LOC across the split files
vs. the original god module.

Mythos chain step 7 of 14. Next: U8 extracts inline tests from all
post-split submodules to integration tests under
`crates/flui-painting/tests/`.
